[svn] The big thing here is recursive extraction.  Find archive files in the thing we're extracting, and extract those when we're done, if the user provides the -r option.  While I was at it:
* --help/--version switches too.
* Throw a more useful error if we get a weird encoding.
* Some other small fix I'm forgetting.

I should definitely add tests for this.  I also really want to step back
and reevaluate the structure of ExtractorApplication.  I don't expect it to
be a shining example of good OO anyway, but it's getting moderately hairy
now, with lots of shared state and side effects and stuff.

--HG--
branch : trunk