[svn] Make ExtractorApplication suck less.  Now the strategies for handling different archive types are out in their own classes, and polymorphism takes care of everything for us.  This is way cleaner.
While I was at it I changed the behavior in the case where an archive
contains one directory that doesn't match the basename.  I now treat that
the same as a bomb.  This can lead to silly directory structures but
ensures that there's no "data" loss nor unexpected results.

--HG--
branch : trunk