[svn] If an archive contains other archives, and the user didn't specify that they wanted recursion, prompt them to find out what they want to do.
This required some changes in the way prompting for one-item archives works
too, because if an archive inside another archive is itself a one-item
archive, things get really weird.  I'm still not really sure what the right
policy is for that.

I'm not wild about this code.  It feels like programming a bunch of special
cases.  I need to figure out a better way to abstract it.  I'm thinking
some kind of Policy class....

--HG--
branch : trunk