[svn] Make dtrx -v print *something*.
paulv wants it to act like tar -v.  I'm not yet convinced this is a good
idea because it makes less sense with dtrx moving things around on you.  It
could do an ls for you, of course, but... why?  But it definitely is weird
to have it be silent.

--HG--
branch : trunk