[svn] Fix various bugs in the recursive extraction.
First, 5.0 introduced a regression where it wouldn't find archives hidden
in subdirectories.  So fix that.

Then recursive extraction would get confused if the archive was a ONE_ENTRY
and we decided to wrap it or just extract it here.  Fix that too: the
extractor keeps track of where it thinks it's extracting to, and then the
handler will change that if necessary.

There are unit tests for all this.

I twiddled some other small thing while I was at it but now I forget what.

--HG--
branch : trunk