Deal with partially extracted tarballs.
If you can only extract part of a tarball -- for example, because it
includes nodes and you're not root -- tar will extract what it can but give
an error code.  So now dtrx will only fail if some subprogram threw an
error code *and* nothing actually got extracted.  Otherwise, we'll show
whatever errors came up, but we'll still assume the extraction was as
successful as it's going to be.

--HG--
branch : trunk