Start printing results ASAP with -l or -t.
Doing -l on a large archive was painful because dtrx would wait until it
had all the listings before it displayed any.  It did this to make sure
that the listing would be successful.

On reconsidering, though, I think it's really unlikely that there'll be a
case where a lister process provides *some* good output for a given file in
the wrong format.  So, now -l will try to get one good line out of the
lister.  If it gets that, then it will immediately start displaying results
as they come in.  On the off chance that it detects an error later on, it
will display an error about that, and then try again with a different
extractor if appropriate.

--HG--
branch : trunk